initial refactoring and organization of file system
