+++ .gitignore
@@ -122,3 +122,5 @@ gradlew
 gradlew.bat
 
 gradle/
+
+.DS_Store
