--- src/main/java/computerPlayer/ComputerPlayer.java
@@ -1,7 +1,0 @@
-package computerPlayer;
-
-import game.Player;
-
-public class ComputerPlayer extends Player {
-
-}
