+++ src/main/java/controller/ChessBoardController.java
@@ -1,0 +1,9 @@
+package controller;
+
+
+// Controller classes that handles user input from ChessPanel and makes the appropriate calls to the board class
+public class ChessBoardController {
+
+
+
+}
