+++ src/main/java/entities/pieces/Bishop.java
@@ -1,8 +1,8 @@
-package pieces;
+package entities.pieces;
 
-public class Queen extends Piece {
+public class Bishop extends Piece {
 
-    public Queen(boolean white, int column, int row) {
+    public Bishop(boolean white, int column, int row) {
         super(white,column, row);
     }
 
@@ -10,7 +10,7 @@ public class Queen extends Piece {
     public boolean canMove(Piece[][] board, int[] start, int[] end) {
         int x = end[0] - start[0];
         int y = end[1] - start[1];
-        if (Math.abs(x) != Math.abs(y) && x != 0 && y != 0) {
+        if (Math.abs(x) != Math.abs(y)) {
             return false;
         }
         if (x > 0 && y > 0) {
@@ -41,43 +41,16 @@ public class Queen extends Piece {
                 }
             }
         }
-        else if (x > 0) {
-            for (int i = 1; i < x; i++) {
-                if (board[start[0] + i][start[1]] != null) {
-                    return false;
-                }
-            }
-        }
-        else if (x < 0) {
-            for (int i = 1; i < Math.abs(x); i++) {
-                if (board[start[0] - i][start[1]] != null) {
-                    return false;
-                }
-            }
-        }
-        else if (y > 0) {
-            for (int i = 1; i < y; i++) {
-                if (board[start[0]][start[1] + i] != null) {
-                    return false;
-                }
-            }
-        }
-        else if (y < 0) {
-            for (int i = 1; i < Math.abs(y); i++) {
-                if (board[start[0]][start[1] - i] != null) {
-                    return false;
-                }
-            }
-        }
         return true;
     }
+
 
     @Override
     public String toString() {
         if (white) {
-            return WHITE_COLOR + "Q" + RESET_COLOR;
+            return WHITE_COLOR + "B" + RESET_COLOR;
         } else {
-            return BLACK_COLOR + "q" + RESET_COLOR;
+            return BLACK_COLOR + "b" + RESET_COLOR;
         }
     }
 }
