+++ src/main/java/entities/pieces/King.java
@@ -1,4 +1,4 @@
-package pieces;
+package entities.pieces;
 
 public class King extends Piece {
