+++ src/main/java/entities/pieces/Knight.java
@@ -1,4 +1,4 @@
-package pieces;
+package entities.pieces;
 
 public class Knight extends Piece {
