--- src/main/java/game/gameplay/HumanComputerGame.java
@@ -1,6 +1,0 @@
-package game.gameplay;
-
-import game.Game;
-
-public class HumanComputerGame extends Game {
-}
