--- src/main/java/game/gameplay/OnlineGame.java
@@ -1,6 +1,0 @@
-package game.gameplay;
-
-import game.Game;
-
-public class OnlineGame extends Game {
-}
