--- src/main/java/game/gameplay/TwoPlayerGame.java
@@ -1,7 +1,0 @@
-package game.gameplay;
-
-import game.Game;
-
-public class TwoPlayerGame extends Game {
-
-}
