+++ src/main/java/useCases/Game.java
@@ -1,4 +1,4 @@
-package game;
+package useCases;
 
 public class Game {
