+++ src/main/java/useCases/Player.java
@@ -1,4 +1,4 @@
-package game;
+package useCases;
 
 public class Player {
